+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - you can now open a dialog to view past status messages and the log output 
   from your scripts (for [Issue #166](https://github.com/pbek/QOwnNotes/issues/166))
     - use the new script command `script.log()` to log to that dialog
+- there now is a new menu entry for exporting the HTML of your note preview
+  (for [Issue #168](https://github.com/pbek/QOwnNotes/issues/168))
 
 ## 16.05.12
 - tags and notes can now be exposed to the QML scripting engine